Add files via upload
Description

This is a simple logger extention for JETI remote control devices.
You can select switchItems you want to log. The switchItem should be 
configured as Center, Prop (not Rev.)
Depending on your device you can have upto 24 log channels in total over
all applications. If you wand to deletea log channel, go to the Item 
select it and then clear (KEY4) the Item. 
It is not possible to register a switchItem twice.
Due to the system design it also makes no sence to sort the log channels.

This application requiers V4.28